SLSTRL1 RGB NoData Handling (#2)
* SLSTRL1 RGB NoData fix

Some of the pixels in input data contain NoData values for one or more
of the RGB bands. These pixels will show up as Red, Green or Blu in the
output granule. To avoid this we set NoData for all three RGB bands in
case one or more of the input bands have an exception flag set for the
pixel.

* Adjust overlap value in config.xml

* revert NoData fix for OLCI L2 product